Set a different build folder for dev, add clean tasks
Before this commit the building regardless of if it was local/dev or for
deploy/distribution was being made to the same folder.

Since we chose to commit/check-in the distribution folder (for archiving
preservation purposes) this would force to check-in many changes into
the repository during the development process.

By having different folders for the different builds, lets us ignore the
constant changes during the development process but still check-in
distribution files intended for deployment.